requirements.txt
pip~=23.0
setuptools~=67.2.0
wheel~=0.38.4
python-dotenv~=0.21.1
requests~=2.28.2
opencv-python~=4.7.0.68
matplotlib~=3.6.3
pillow~=9.4.0
numpy~=1.24.2
